+++ custom_components/rfplayer/number.py
@@ -1,53 +1,61 @@
 """Support for Rfplayer number."""
 import logging
 
-from homeassistant.components.number import NumberEntity
-from homeassistant.core import callback
-from homeassistant.helpers.entity import EntityCategory
+from homeassistant.components.number import NumberMode, RestoreNumber
+from homeassistant.config_entries import ConfigEntry
+from homeassistant.core import HomeAssistant, callback
+from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from . import RfplayerDevice
-from .const import DOMAIN, RFPLAYER_PROTOCOL
+from . import EVENT_KEY_COMMAND, RfplayerDevice
+from .const import DATA_ENTITY_LOOKUP, DOMAIN, EVENT_KEY_ID, RFPLAYER_PROTOCOL
 
 _LOGGER = logging.getLogger(__name__)
 
 
-async def async_setup_entry(hass, entry, async_add_entities):
+async def async_setup_entry(
+    hass: HomeAssistant, entry: ConfigEntry, async_add_entities: AddEntitiesCallback
+) -> None:
     """Set up the Rfplayer platform."""
     _LOGGER.debug("Add jamming number entity")
     async_add_entities([RfplayerJammingNumber()])
 
 
-class RfplayerJammingNumber(RfplayerDevice, NumberEntity):
+class RfplayerJammingNumber(RfplayerDevice, RestoreNumber):
     """Representation of a Rfplayer jamming number setting entity."""
 
-    def __init__(self):
-        """Init the number rfplayer entity."""
-        self._attr_native_min_value = 0
-        self._attr_native_max_value = 10
-        self._attr_native_mode = "slider"
-        self._attr_native_entity_category = EntityCategory.CONFIG
-        super().__init__("JAMMING", device_id=0, name="Jamming detection level")
+    _attr_native_min_value = 0
+    _attr_native_max_value = 10
+    _attr_mode = NumberMode.SLIDER
 
-    async def async_added_to_hass(self):
+    def __init__(self) -> None:
+        """Init the number rfplayer entity."""
+        self._state: int | None = None
+        super().__init__("JAMMING")
+
+    async def async_added_to_hass(self) -> None:
         """Restore RFPlayer device state."""
         await super().async_added_to_hass()
+
+        self.hass.data[DOMAIN][DATA_ENTITY_LOOKUP][EVENT_KEY_COMMAND][
+            self._initial_event[EVENT_KEY_ID]
+        ] = self.entity_id
+
         if self._event is None:
             old_state = await self.async_get_last_state()
             if old_state is not None:
-                self._state = old_state.state
+                self._state = int(old_state.state)
 
     @callback
     def _handle_event(self, event):
-        self._state = int(event["value"])
+        self._state = int(event["command"])
 
     @property
-    def native_value(self):
+    def value(self):
         """Return the current setting."""
         return self._state
 
-    async def async_set_native_value(self, value) -> None:
+    async def async_set_native_value(self, value: float) -> None:
         """Update the current value."""
         rfplayer = self.hass.data[DOMAIN][RFPLAYER_PROTOCOL]
         await rfplayer.send_command_ack(command=int(value), protocol=self._protocol)
-        self._state = value
-        self.async_write_ha_state()
+        self._state = int(value)