+++ custom_components/rfplayer/rflib/rfpprotocol.py
@@ -1,16 +1,18 @@
 """Asyncio protocol implementation of RFplayer."""
 
 import asyncio
+from collections.abc import Callable, Coroutine, Sequence
 from datetime import timedelta
 from fnmatch import fnmatchcase
 from functools import partial
 import logging
-from typing import Any, Callable, Coroutine, Optional, Sequence, Tuple, Type
+from typing import Any, Optional, cast
 
 from serial_asyncio import create_serial_connection
 
 from .rfpparser import (
     PacketType,
+    RfPlayerException,
     decode_packet,
     encode_packet,
     packet_events,
@@ -25,27 +27,22 @@ TIMEOUT = timedelta(seconds=5)
 class ProtocolBase(asyncio.Protocol):
     """Manage low level rfplayer protocol."""
 
-    transport = None  # type: asyncio.BaseTransport
-
     def __init__(
         self,
-        loop: Optional[asyncio.AbstractEventLoop] = None,
         disconnect_callback: Optional[Callable[[Optional[Exception]], None]] = None,
         **kwargs: Any,
     ) -> None:
         """Initialize class."""
-        if loop:
-            self.loop = loop
-        else:
-            self.loop = asyncio.get_event_loop()
+        self.transport: asyncio.WriteTransport | None = None
         self.packet = ""
         self.buffer = ""
-        self.packet_callback = None  # type: Optional[Callable[[PacketType], None]]
+        self.packet_callback: Callable[[PacketType], None] | None = None
         self.disconnect_callback = disconnect_callback
 
     def connection_made(self, transport: asyncio.BaseTransport) -> None:
         """Just logging for now."""
-        self.transport = transport
+
+        self.transport = cast(asyncio.WriteTransport, transport)
         log.debug("connected")
         self.send_raw_packet("ZIA++HELLO")
         self.send_raw_packet("ZIA++RECEIVER + *")
@@ -75,13 +72,14 @@ class ProtocolBase(asyncio.Protocol):
 
     def handle_raw_packet(self, raw_packet: str) -> None:
         """Handle one raw incoming packet."""
-        raise NotImplementedError()
+        raise NotImplementedError
 
     def send_raw_packet(self, packet: str) -> None:
         """Encode and put packet string onto write buffer."""
         data = bytes(packet + "\n\r", "utf-8")
         log.debug("writing data: %s", repr(data))
-        self.transport.write(data)  # type: ignore
+        assert self.transport is not None
+        self.transport.write(data)
 
     def connection_lost(self, exc: Optional[Exception]) -> None:
         """Log when connection is closed, if needed call callback."""
@@ -117,7 +115,7 @@ class PacketHandling(ProtocolBase):
         packets = []
         try:
             packets = decode_packet(raw_packet)
-        except BaseException:
+        except RfPlayerException:
             log.exception("failed to parse packet data: %s", raw_packet)
 
         if packets:
@@ -138,11 +136,11 @@ class PacketHandling(ProtocolBase):
             # forward to callback
             self.packet_callback(packet)
         else:
-            print("packet", packet)
+            log.debug("packet %s", packet)
 
     def handle_response_packet(self, packet: PacketType) -> None:
         """Handle response packet."""
-        raise NotImplementedError()
+        raise NotImplementedError
 
     def send_packet(self, fields: PacketType) -> None:
         """Concat fields and send packet to gateway."""
@@ -153,21 +151,17 @@ class PacketHandling(ProtocolBase):
         self,
         protocol: str,
         command: str,
-        device_address: str = None,
-        device_id: str = None,
+        device_address: str | None = None,
+        device_id: str | None = None,
     ) -> None:
         """Send device command to rfplayer gateway."""
         if device_id is not None:
-            if protocol == "EDISIOFRAME" :
-                self.send_raw_packet(f"ZIA++{protocol} {device_id}")
-            else :
-                self.send_raw_packet(f"ZIA++{command} {protocol} ID {device_id}")
+            self.send_raw_packet(f"ZIA++{command} {protocol} ID {device_id}")
         elif device_address is not None:
             self.send_raw_packet(f"ZIA++{command} {protocol} {device_address}")
-        elif protocol == "EDISIOFRAME":
-            self.send_raw_packet(f"ZIA++{command}")
         else:
             self.send_raw_packet(f"ZIA++{protocol} {command}")
+
 
 class CommandSerialization(PacketHandling):
     """Logic for ensuring asynchronous commands are sent in order."""
@@ -185,6 +179,7 @@ class CommandSerialization(PacketHandling):
             self.packet_callback = packet_callback
         self._event = asyncio.Event()
         self._lock = asyncio.Lock()
+        self._last_ack: PacketType | None = None
 
     def handle_response_packet(self, packet: PacketType) -> None:
         """Handle response packet."""
@@ -196,8 +191,8 @@ class CommandSerialization(PacketHandling):
         self,
         protocol: str,
         command: str,
-        device_address: str = None,
-        device_id: str = None,
+        device_address: str | None = None,
+        device_id: str | None = None,
     ) -> bool:
         """Send command, wait for gateway to repond."""
         async with self._lock:
@@ -266,7 +261,7 @@ class EventHandling(PacketHandling):
             if event.get("unit"):
                 string += " {unit}"
 
-        print(string.format(**event))
+        log.debug(string.format(**event))
 
     def handle_packet(self, packet: PacketType) -> None:
         """Apply event specific handling and pass on to packet handling."""
@@ -276,10 +271,7 @@ class EventHandling(PacketHandling):
     def ignore_event(self, event_id: str) -> bool:
         """Verify event id against list of events to ignore."""
         log.debug("ignore_event")
-        for ignore in self.ignore:
-            if fnmatchcase(event_id, ignore):
-                return True
-        return False
+        return any(fnmatchcase(event_id, ignore) for ignore in self.ignore)
 
 
 class RfplayerProtocol(CommandSerialization, EventHandling):
@@ -289,19 +281,18 @@ class RfplayerProtocol(CommandSerialization, EventHandling):
 def create_rfplayer_connection(
     port: str,
     baud: int = 115200,
-    protocol: Type[ProtocolBase] = RfplayerProtocol,
     packet_callback: Optional[Callable[[PacketType], None]] = None,
     event_callback: Optional[Callable[[PacketType], None]] = None,
     disconnect_callback: Optional[Callable[[Optional[Exception]], None]] = None,
     ignore: Optional[Sequence[str]] = None,
     loop: Optional[asyncio.AbstractEventLoop] = None,
-) -> "Coroutine[Any, Any, Tuple[asyncio.BaseTransport, ProtocolBase]]":
+) -> "Coroutine[Any, Any, tuple[asyncio.BaseTransport, ProtocolBase]]":
     """Create Rflink manager class, returns transport coroutine."""
     if loop is None:
         loop = asyncio.get_event_loop()
     # use default protocol if not specified
     protocol_factory = partial(
-        protocol,
+        RfplayerProtocol,
         loop=loop,
         packet_callback=packet_callback,
         event_callback=event_callback,